Moved helper to RefactoringHelpersService.
