+++ src/Features/Core/Portable/CodeRefactoringHelpers.cs
@@ -12,99 +12,6 @@ namespace Microsoft.CodeAnalysis
 {
     internal static class CodeRefactoringHelpers
     {
-        /// <summary>
-        /// <para>
-        /// Returns an instance of <typeparamref name="TSyntaxNode"/> for refactoring given specified selection in document or null
-        /// if no such instance exists.
-        /// </para>
-        /// <para>
-        /// A <typeparamref name="TSyntaxNode"/> instance is returned if:
-        /// - Selection is zero-width and inside/touching a Token with direct parent of type <typeparamref name="TSyntaxNode"/>.
-        /// - Selection is zero-width and touching a Token whose ancestor ends/starts precisely on current selection .
-        /// - Token whose direct parent of type <typeparamref name="TSyntaxNode"/> is selected.
-        /// - Whole node of a type <typeparamref name="TSyntaxNode"/> is selected.
-        /// </para>
-        /// <para>
-        /// Note: this function strips all whitespace from both the beginning and the end of given <paramref name="selection"/>.
-        /// The stripped version is then used to determine relevant <see cref="SyntaxNode"/>. It also handles incomplete selections
-        /// of tokens gracefully.
-        /// </para>
-        /// </summary>
-        public static Task<TSyntaxNode> TryGetSelectedNodeAsync<TSyntaxNode>(
-            Document document, TextSpan selection, CancellationToken cancellationToken) where TSyntaxNode : SyntaxNode
-        {
-            return TryGetSelectedNodeAsync<TSyntaxNode>(document, selection, Functions<SyntaxNode>.Identity, cancellationToken);
-        }
-
-        public static async Task<TSyntaxNode> TryGetSelectedNodeAsync<TSyntaxNode>(
-            Document document, TextSpan selection, Func<SyntaxNode, SyntaxNode> extractNode, CancellationToken cancellationToken) where TSyntaxNode : SyntaxNode
-        {
-            return await TryGetSelectedNodeAsync(document, selection, extractNode, n => n is TSyntaxNode, cancellationToken).ConfigureAwait(false) as TSyntaxNode;
-        }
-
-        public static async Task<SyntaxNode> TryGetSelectedNodeAsync(
-            Document document, TextSpan selection, Func<SyntaxNode, SyntaxNode> extractNode, Predicate<SyntaxNode> predicate, CancellationToken cancellationToken)
-        {
-            var root = await document.GetSyntaxRootAsync(cancellationToken).ConfigureAwait(false);
-            var selectionStripped = await GetStrippedTextSpan(document, selection, cancellationToken).ConfigureAwait(false);
-
-            // Handle selections
-            var node = root.FindNode(selectionStripped, getInnermostNodeForTie: true);
-            SyntaxNode prevNode;
-            do
-            {
-                var extrNode = extractNode(node);
-                if (extrNode != default && predicate(extrNode))
-                {
-                    return extrNode;
-                }
-
-                prevNode = node;
-                node = node.Parent;
-
-            }
-            while (node != null && prevNode.FullWidth() == node.FullWidth());
-
-            // only consider what is direct selection touching when selection is empty 
-            // prevents `[|C|] methodName(){}` from registering as relevant for method Node
-            if (!selection.IsEmpty)
-            {
-                return default;
-            }
-
-            var tokenToLeft = await root.SyntaxTree.GetTouchingTokenToLeftAsync(selectionStripped.Start, cancellationToken).ConfigureAwait(false);
-            var leftNode = tokenToLeft.Parent;
-            do
-            {
-                // either touches a Token which parent is `TSyntaxNode` or is whose ancestor's span ends on selection
-                var extrNode = extractNode(leftNode);
-                if (extrNode != default && predicate(extrNode))
-                {
-                    return extrNode;
-                }
-
-                leftNode = leftNode?.Parent;
-            }
-            while (leftNode != null && leftNode.Span.End == selection.Start);
-
-            var tokenToRight = await root.SyntaxTree.GetTouchingTokenToRightOrInAsync(selectionStripped.Start, cancellationToken).ConfigureAwait(false);
-            var rightNode = tokenToRight.Parent;
-            do
-            {
-                // either touches a Token which parent is `TSyntaxNode` or is whose ancestor's span starts on selection
-                var extrNode = extractNode(rightNode);
-                if (extrNode != default && predicate(extrNode))
-                {
-                    return extrNode;
-                }
-
-                rightNode = rightNode?.Parent;
-            }
-            while (rightNode != null && rightNode.Span.Start == selection.Start);
-
-            return default;
-        }
-
         public static Task<bool> RefactoringSelectionIsValidAsync(
             Document document, TextSpan selection, SyntaxNode node, CancellationToken cancellation)
         {
@@ -237,7 +144,7 @@ namespace Microsoft.CodeAnalysis
         /// Returns unchanged <paramref name="span"/> in case <see cref="TextSpan.IsEmpty"/>.
         /// Returns empty Span with original <see cref="TextSpan.Start"/> in case it contains only whitespace.
         /// </remarks>
-        private static async Task<TextSpan> GetStrippedTextSpan(Document document, TextSpan span, CancellationToken cancellationToken)
+        public static async Task<TextSpan> GetStrippedTextSpan(Document document, TextSpan span, CancellationToken cancellationToken)
         {
             if (span.IsEmpty)
             {
